+++ mpqc/src/lib/util/group/memproc.h
@@ -46,7 +46,7 @@ class ProcMemoryGrp: public MemoryGrp {
     ProcMemoryGrp(const Ref<KeyVal>&);
     ~ProcMemoryGrp();
 
-    void set_localsize(int);
+    void set_localsize(size_t);
     void *localdata();
 
     void *obtain_readwrite(distsize_t offset, int size);
